// src/transactions/dto/transaction.dto.ts
export class TransactionDto {
  userId?: string;
  merchantId?: string;
  clientId?: string;
  retailer?: any;
  clientSecret?: string;
  customerMsisdn?: string;
  recipientMsisdn?: string;
  senderMsisdn?: string;
  channel?: string;
  amount: any;
  primaryCallbackUrl?: string;
  clientReference?: string;
  description?: string;
  merchantName?: string;
  transType: string;
  serviceName?: string;
  customerName?: string;
  customerEmail?: string;
  merchantReference?: string;
  transId: string;
  originalAmount: string;
  amountPaid: string;
  charge: string;
  network: string;
  mobileNumber: string;
  recipientName: string;
  recipientNumber: string;
  senderNumber: string;
  product: string;
  transDescription: string;
  voucher: string;
  serviceStatus: string;
  transMessage: string;
  serviceTransId: string;
  transStatus: string;
  paymentStatus: string;
  callbackReceipt: string;
  commentary: string;
  callbackUrl: string;
  readonly price: string;
  recipient_number: string;
  sender: string;
  award?: string;
  apikey: string;
  orderID: string;
  token?: string;
  notifyURL?: string;
  notificationFormat?: string;
  destinationAddress?: string;
  criteria?: string;
  Status?: string;
  Transactionid?: string;
  transactionid?: string;
  Message?: string;
  MerchantReference?: string;
  CallbackURL?: string;
  currentBalance: string;
  balanceBefore: number;
  balanceAfter: number;
  recipient: string;
  trxn: string;
  fee: any;
  serviceCode: string;
  serviceStatusCode?: string;
  statusNarration?: string;
  redirectURL?: string;
}